+++ public/iubenda-config.js
@@ -1,0 +1,9 @@
+window._iub = window._iub || [];
+window._iub.csConfiguration = {
+  siteId: 4122415,
+  floatingPreferencesButtonDisplay: false,
+  cookiePolicyId: 16849142,
+  lang: 'en',
+  storage: { useSiteId: true },
+};
+